chore: Update scalameta to newest version
