+++ metals-bench/src/main/scala/scala/meta/bench/ClasspathIndexingBench.scala
@@ -1,4 +1,4 @@
-package bench
+package scala.meta.bench
 
 import java.nio.file.Path
 import java.util.concurrent.TimeUnit
