+++ metals-bench/src/main/scala/scala/meta/bench/MainBench.scala
@@ -1,4 +1,5 @@
-package bench
+package scala.meta.bench
+import scala.meta.bench.ClasspathSymbolsBench
 
 object MainBench {
   def main(args: Array[String]): Unit = {
