+++ metals-bench/src/main/scala/scala/meta/bench/MetalsBench.scala
@@ -1,4 +1,4 @@
-package bench
+package scala.meta.bench
 
 import scala.reflect.internal.util.BatchSourceFile
 import scala.reflect.io.VirtualFile
