+++ metals-bench/src/main/scala/scala/meta/bench/PcBenchmark.scala
@@ -1,4 +1,4 @@
-package bench
+package scala.meta.bench
 
 import java.nio.file.Path
